Changed matrixes to long to prevent overflow

+++ src/matrixEngine.java
@@ -2,8 +2,8 @@ import java.util.Random;
 
 public class matrixEngine {
     public matrixResult GenerateMatrixes(){
-        int[][] matrix1 = new int[1000][1000];
-        int[][] matrix2 = new int[1000][1000];
+        long[][] matrix1 = new long[1000][1000];
+        long[][] matrix2 = new long[1000][1000];
 
         fillMatrix(matrix1);
         fillMatrix(matrix2);
@@ -11,7 +11,7 @@ public class matrixEngine {
         return new matrixResult(matrix1, matrix2);
     }
 
-    public void fillMatrix(int[][] matrix){
+    public void fillMatrix(long[][] matrix){
         Random random = new Random();
 
         for (int i = 0; i < matrix.length; i++) {
@@ -21,8 +21,8 @@ public class matrixEngine {
         }
     }
 
-    public int[][] multiplyMatrices(int[][] matrix1, int[][] matrix2) {
-        int[][] resultMatrix = new int[1000][1000];
+    public long[][] multiplyMatrices(long[][] matrix1, long[][] matrix2) {
+        long[][] resultMatrix = new long[1000][1000];
         for (int i = 0; i < matrix1.length; i++) {
             for (int j = 0; j < matrix2[0].length; j++) {
                 for (int k = 0; k < matrix2.length; k++) {
