+++ src/matrixResult.java
@@ -1,8 +1,8 @@
 public class matrixResult {
-    int[][] matrix1;
-    int[][] matrix2;
+    long[][] matrix1;
+    long[][] matrix2;
 
-    public matrixResult(int[][] matrix1, int[][] matrix2){
+    public matrixResult(long[][] matrix1, long[][] matrix2){
         this.matrix1 = matrix1;
         this.matrix2 = matrix2;
     }
